updatedistilbert with text_classification (#529)

+++ quickstart/language_modeling/pytorch/distilbert_base/inference/cpu/run_multi_instance_throughput.sh
@@ -55,25 +55,29 @@ mode="jit"
 ARGS="$ARGS --jit_mode"
 echo "### running with jit mode"
 
-
-CORES=`lscpu | grep Core | awk '{print $4}'`
-BATCH_SIZE=${BATCH_SIZE:-`expr 4 \* $CORES`}
-FINETUNED_MODEL=${FINETUNED_MODEL:-"distilbert-base-uncased-distilled-squad"}
 if [ -z "${OUTPUT_DIR}" ]; then
   echo "The required environment variable OUTPUT_DIR has not been set, please create the output path and set it to OUTPUT_DIR"
   exit 1
 fi
-EVAL_SCRIPT=${EVAL_SCRIPT:-"./transformers/examples/pytorch/question-answering/run_qa.py"}
+if [ -z "${SEQUENCE_LENGTH}" ]; then
+  echo "The required environment variable SEQUENCE_LENGTH has not been set, please set the seq_length before running"
+  exit 1
+fi
+
+CORES=`lscpu | grep Core | awk '{print $4}'`
+BATCH_SIZE=${BATCH_SIZE:-`expr 4 \* $CORES`}
+FINETUNED_MODEL=${FINETUNED_MODEL:-"distilbert-base-uncased-finetuned-sst-2-english"}
+
+EVAL_SCRIPT=${EVAL_SCRIPT:-"./transformers/examples/pytorch/text-classification/run_glue.py"}
 WORK_SPACE=${WORK_SPACE:-${OUTPUT_DIR}}
 
 rm -rf ${OUTPUT_DIR}/throughput_log*
 python -m intel_extension_for_pytorch.cpu.launch --ninstance 1 --node_id 0  --enable_jemalloc --log_path=${OUTPUT_DIR} --log_file_prefix="./throughput_log_${path}_${precision}_${mode}" \
   ${EVAL_SCRIPT} $ARGS \
   --model_name_or_path   ${FINETUNED_MODEL} \
-  --dataset_name squad \
+  --task_name sst2 \
   --do_eval \
-  --max_seq_length 128 \
-  --doc_stride 64 \
+  --max_seq_length ${SEQUENCE_LENGTH} \
   --output_dir ./tmp \
   --per_device_eval_batch_size $BATCH_SIZE \
